model cleanup & update references
