--- src/frontend/src/models/ClusterSummary.ts
@@ -1,9 +1,0 @@
-export class ClusterSummary {
-    name:string = ''
-    version:string = ''
-    grade:string = ''
-    score:number = 0
-    nodes:number = 0
-    namespaces:number = 0
-    pods:number = 0
-}
