+++ src/frontend/src/models/Collection.ts
@@ -1,0 +1,15 @@
+/// <summary>
+/// Collection type and name of a check.
+/// (eg: namespace: default, resource group: common).
+/// </summary>
+export class Collection {
+  /// <summary>
+  /// Type of the collection.
+  /// </summary>
+  public type: string = '';
+
+  /// <summary>
+  /// Name of the collection.
+  /// </summary>
+  public name: string = '';
+}
