+++ src/frontend/src/models/ComponentDiff.ts
@@ -1,0 +1,11 @@
+import { InfrastructureComponentSummary } from '@/models/';
+
+export class InfrastructureComponentDiff {
+
+  ///Components of first summary.
+  summary1: InfrastructureComponentSummary = new InfrastructureComponentSummary();
+
+  ///Components of second summary.
+  summary2: InfrastructureComponentSummary = new InfrastructureComponentSummary();
+}
+
