--- src/frontend/src/models/ImageScan.ts
@@ -1,7 +1,0 @@
-import { ContainerImageScan } from '.';
-import { ImageScanGroup } from './ImageScanGroup'
-
-export class ImageScan {
-    scans: ContainerImageScan[] = []
-    groups: ImageScanGroup[] = []
-}
