+++ src/frontend/src/models/ImageScanDetailModel.ts
@@ -1,0 +1,8 @@
+import { TargetGroup } from '@/models/';
+
+export class ImageScanDetailModel {
+  image: string = "";
+  scanResult: string = "";
+  description: string = "";
+  targets: TargetGroup[] = []
+}
