--- src/frontend/src/models/ImageScanGroup.ts
@@ -1,7 +1,0 @@
-export class ImageScanGroup {
-    constructor(
-        public count: number, 
-        public title: string, 
-        public description: string
-    ){ }
-}
