--- src/frontend/src/models/NamespaceCounters.ts
@@ -1,7 +1,0 @@
-export class NamespaceCounters {
-  constructor(
-    public failing: number = 0,
-    public passing: number = 0,
-    public nodata: number = 0
-  ) {}
-}
