+++ src/frontend/src/models/Resource.ts
@@ -1,0 +1,17 @@
+/// Resource of a check.
+export class Resource {
+  /// <summary>
+  /// Id of the resource.
+  /// </summary>
+  public id: string = '';
+
+  /// <summary>
+  /// Name of the resource.
+  /// </summary>
+  public name: string = '';
+
+  /// <summary>
+  /// Type of the resource.
+  /// </summary>
+  public type: string = '';
+}
