--- src/frontend/src/models/ResultSummary.ts
@@ -1,7 +1,0 @@
-export class ResultSummary {
-    resultName: string = ''
-    Successes:number = 0
-    Warnings:number = 0 
-    Errors:number = 0
-    NoDatas:number = 0
-}
