--- src/frontend/src/models/ScanTarget.ts
@@ -1,6 +1,0 @@
-import { Vulnerability } from "./InfrastructureOverview";
-
-export class ScanTarget {
-  target: string = "";
-  vulnerabilities: Vulnerability[] = [];
-}
