+++ src/frontend/src/models/ScoreHistory.ts
@@ -1,0 +1,5 @@
+
+/// Represents the score of component at a given date.
+export class ScoreHistoryItem {
+  constructor(public recordedAt: Date, public score: number) { }
+}
