+++ src/frontend/src/models/TargetGroup.ts
@@ -1,0 +1,6 @@
+import { VulnerabilityGroup } from '@/models/';
+
+export class TargetGroup {
+  vulnerabilities: VulnerabilityGroup[] = [];
+  constructor(public target: string) { }
+}
