+++ src/frontend/src/models/Vulnerability.ts
@@ -1,0 +1,10 @@
+export class Vulnerability {
+  VulnerabilityID: string = "";
+  PkgName: string = "";
+  InstalledVersion: string = "";
+  FixedVersion: string = "";
+  Title: string = ""; //
+  Description: string = "";
+  Severity: string = "";
+  References: string[] = [];
+}
