--- src/frontend/src/models/VulnerabilityCounter.ts
@@ -1,4 +1,0 @@
-export class VulnerabilityCounter {
-	severity:string = ''
-	count:number = 0
-}
