+++ src/frontend/src/models/VulnerabilityGroup.ts
@@ -1,0 +1,8 @@
+import { Vulnerability } from '@/models/';
+
+export class VulnerabilityGroup {
+  Count: number = 0;
+  CVEs: Vulnerability[] = [];
+  Order: number = 0;
+  constructor(public Severity: string) { }
+}
